Moved Pricing APIs to their own namespace
